requirements.txt
openpyxl==3.1.2
sympy==1.12
psycopg2-binary==2.9.9